+++ validators/signupValidator.js
@@ -11,11 +11,11 @@ const confirmPassword = (password, { req }) => {
  * Return a Promise or throw a error
  * https://express-validator.github.io/docs/guides/customizing/#implementing-a-custom-validator
  */
-const validateFullname = async (fullname) => {
+const fullnameValidator = async (fullname) => {
   const regex = new RegExp("(^[A-Za-z]{3,16})([ ]{0,1})([A-Za-z]{3,16})");
 };
 
-const validateUsername = async (username) => {
+const usernameValidator = async (username) => {
   // Test against regex
   // Need to make sure user name is not taken
   const regex = new RegExp("^[a-zA-Z\\-\\_]{3,10}$");
@@ -44,7 +44,7 @@ const accountSchema = {
   username: {
     trim: true,
     custom: {
-      options: validateUsername,
+      options: usernameValidator,
     },
     errorMessage:
       "Username must be between 3 and 10 letters long. Dashes and underscores are the only acceptable symbols.",